Move custom editor files to the Editor folder.
