+++ Packages/HVRBasisAvatarComms/Scripts/Editor/AcquisitionAssistEditor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: fe8fa0a2e4614eb08c07ecc15da2e1d9
+timeCreated: 1729329910
