+++ Packages/HVRBasisAvatarComms/Scripts/Editor/StreamedAvatarFeatureAssistEditor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 8789ec4477d449f5a46035419503a8fd
+timeCreated: 1729329963
